requirements.txt
coverage==7.9.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pycryptodome==3.23.0
Pygments==2.19.2
pytest==8.4.1

